+++ project/Routes/console.php
@@ -5,12 +5,14 @@ use Project\Console\Commands\SyncUiDocsGettingStartedCommand;
 use Project\Console\Commands\SyncUiDocsHomeMainCommand;
 use Project\Console\Commands\SyncUiDocsNavigationCommand;
 use Project\Console\Commands\SetupWebBlocksUiDocsSiteCommand;
+use Project\Console\Commands\WebBlocksUiLocalResolverCommand;
 use Project\Console\Commands\WebBlocksUiImportCommand;
 
 require_once base_path('project/Console/Commands/SyncUiDocsNavigationCommand.php');
 require_once base_path('project/Console/Commands/SyncUiDocsGettingStartedCommand.php');
 require_once base_path('project/Console/Commands/SyncUiDocsHomeMainCommand.php');
 require_once base_path('project/Console/Commands/SetupWebBlocksUiDocsSiteCommand.php');
+require_once base_path('project/Console/Commands/WebBlocksUiLocalResolverCommand.php');
 require_once base_path('project/Console/Commands/WebBlocksUiImportCommand.php');
 
 ArtisanApplication::starting(function ($artisan): void {
@@ -19,6 +21,7 @@ ArtisanApplication::starting(function ($artisan): void {
         SyncUiDocsGettingStartedCommand::class,
         SyncUiDocsHomeMainCommand::class,
         SetupWebBlocksUiDocsSiteCommand::class,
+        WebBlocksUiLocalResolverCommand::class,
         WebBlocksUiImportCommand::class,
     ]);
 });